pr feedback and add tests

+++ src/windows/service/exe/WslCoreGuestNetworkService.h
@@ -64,10 +64,10 @@ private:
     _Requires_lock_held_(m_dataLock)
     bool IsPortInGuestEphemeralRange(uint16_t PortNumber) const noexcept;
 
-    uint32_t ComputeHostEphemeralPortCap(int Protocol) const noexcept;
+    uint16_t ComputeHostEphemeralPortCap(int Protocol) const noexcept;
 
     _Requires_lock_held_(m_dataLock)
-    uint32_t ComputeHostEphemeralOverlap(int Protocol) const noexcept;
+    uint16_t ComputeHostEphemeralOverlap(int Protocol) const noexcept;
 
     static std::pair<uint16_t, uint16_t> QueryHostEphemeralPortRange(LPCWSTR WmiClassName) noexcept;
 
@@ -83,14 +83,15 @@ private:
     _Guarded_by_(m_dataLock) std::map<std::pair<HCN_PORT_PROTOCOL, USHORT>, HcnPortReservation> m_reservedPorts;
     _Guarded_by_(m_dataLock) HCN_PORT_RANGE_RESERVATION m_reservedPortRange {};
 
-    // Host ephemeral port ranges can change. They are queried once at startup, if a change occurs, the service will need to be restarted.
+    // Host ephemeral port ranges can change. They are queried once at startup, if a change occurs, the service will need to be
+    // restarted. Note: The host ephemeral range will be the same for both IPv4 and IPv6, but can be different for TCP and UDP.
     std::pair<uint16_t, uint16_t> m_hostTcpEphemeralPortRange{};
     std::pair<uint16_t, uint16_t> m_hostUdpEphemeralPortRange{};
 
-    uint32_t m_hostTcpEphemeralPortCap{};
-    uint32_t m_hostUdpEphemeralPortCap{};
+    uint16_t m_hostTcpEphemeralPortCap{};
+    uint16_t m_hostUdpEphemeralPortCap{};
 
-    _Guarded_by_(m_dataLock) uint32_t m_hostTcpEphemeralPortsInUse {};
-    _Guarded_by_(m_dataLock) uint32_t m_hostUdpEphemeralPortsInUse {};
+    _Guarded_by_(m_dataLock) uint16_t m_hostTcpEphemeralPortsInUse {};
+    _Guarded_by_(m_dataLock) uint16_t m_hostUdpEphemeralPortsInUse {};
 };
 } // namespace wsl::core::networking
